extract AcceptRejectTransaction out of AcceptPreauthorizedConversationsController so methods can be used in during user payment (PreauthorizeTransactionsController#initiated)
