+++ app/controllers/concerns/accept_reject_transaction.rb
@@ -1,0 +1,23 @@
+module AcceptRejectTransaction
+  extend ActiveSupport::Concern
+
+  def accept_tx(community_id, tx_id, message, sender_id)
+    TransactionService::Transaction.complete_preauthorization(community_id: community_id,
+                                                              transaction_id: tx_id,
+                                                              message: message,
+                                                              sender_id: sender_id)
+      .maybe()
+      .map { |_| {flow: :accept, success: true}}
+      .or_else({flow: :accept, success: false})
+  end
+
+  def reject_tx(community_id, tx_id, message, sender_id)
+    TransactionService::Transaction.reject(community_id: community_id,
+                                           transaction_id: tx_id,
+                                           message: message,
+                                           sender_id: sender_id)
+      .maybe()
+      .map { |_| {flow: :reject, success: true}}
+      .or_else({flow: :reject, success: false})
+  end
+end
